Transitioning to Reusable Components
In this commit, we are making a strategic shift towards enhancing the application's scalability and maintainability. We are currently in the process of converting various elements within our forms into reusable components. This includes the creation of an `InputField` component to standardize input fields across different forms such as Login and Signup. The goal is to streamline the codebase, making it cleaner and more efficient by reducing redundancy and improving component reusability. Further adjustments and optimizations will follow in subsequent commits.